Don't change invalid baseURLs from loadData or loadHTMLString to about:blank https://bugs.webkit.org/show_bug.cgi?id=228717 <rdar://79909343>
Reviewed by Chris Dumez.

Source/WebKit:

r278737 fixed a compatibility bug from r272469 but introduced a crash when the base URL does not successfully parse.
r279450 fixed that crash but introduced a new behavior change that changes the URL to about:blank.  This didn't seem
to change anything because window.location is about:blank with and without the change, but WKWebView.URL did change,
which broke at least one app.  This undoes the functionality change from r279450 and keeps the crash fix, which are
both covered by API tests.  I also manually verified this fixes the app in the radar.

* WebProcess/WebPage/WebPage.cpp:
(WebKit::WebPage::loadData):

Tools:

* TestWebKitAPI/Tests/WebKitCocoa/LoadAlternateHTMLString.mm:
(TEST):



git-svn-id: http://svn.webkit.org/repository/webkit/trunk@280562 268f45cc-cd09-0410-ab3c-d52691b4dbfc