VoiceOver is not announcing the state as selected/unselected when an item is selected in a grid. https://bugs.webkit.org/show_bug.cgi?id=228840 <rdar://78225826>
Reviewed by Chris Fleizach.

Source/WebCore:

Test: accessibility/selected-state-changed-notifications.html.

On iOS, a table/grid cell element is not exposed to AX clients. Instead
the content of the cell is. This patch ensures that if the cell element
is selected, the AX objects that represent the content inside the cell
will inherit the selected state from the cell.

* accessibility/ios/WebAccessibilityObjectWrapperIOS.mm:
(-[WebAccessibilityObjectWrapper _accessibilityTraitsFromAncestors]):

LayoutTests:

* accessibility/selected-state-changed-notifications.html:
* platform/ios/TestExpectations:
* platform/ios/accessibility/selected-state-changed-notifications-expected.txt: Added.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@280705 268f45cc-cd09-0410-ab3c-d52691b4dbfc