Allow custom IPC::Attachment messaging in ConnectionUnix.cpp https://bugs.webkit.org/show_bug.cgi?id=227740
Reviewed by Žan Doberšek.

In order to share AHardwareBuffer handles via IPC we need to use the AHardwareBuffer_sendHandleToUnixSocket()
and AHardwareBuffer_recvHandleFromUnixSocket() functions. This is not currently compatible with the Attachment
API used in UnixConnection, as those functions use custom messaging into the socked fd.

This patch adds a CustomWriterType Attachment that allows custom messaging going into the socket fd.

* Platform/IPC/unix/ConnectionUnix.cpp:


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@280769 268f45cc-cd09-0410-ab3c-d52691b4dbfc