-Warray-bounds, -Wstringop-truncation warnings in Packed.h https://bugs.webkit.org/show_bug.cgi?id=226557 <rdar://problem/79103658>
Patch by Michael Catanzaro <mcatanzaro@gnome.org> on 2021-08-09
Reviewed by Adrian Perez de Castro.

Suppressing the -Warray-bounds warning introduces the -Wstringop-truncation warning, which
cannot be suppressed with pragmas. The only solution I found is -Wno-stringop-truncation. I
could apply it globally, but this warning does not have many false positives, so I think
it's better to suppress the warning only for WebSocket.cpp. Unfortunately we have to remove
it from the unified build in order to add custom flags, but it seems worth it to avoid
suppressing the warning globally.

* CMakeLists.txt:

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@280778 268f45cc-cd09-0410-ab3c-d52691b4dbfc